requirements.txt
asgiref==3.5.2
autopep8==1.6.0
backports.zoneinfo==0.2.1
certifi==2022.6.15
charset-normalizer==2.1.0
Django==4.0.6
django-cors-headers==3.13.0
djangorestframework==3.13.1
idna==3.3
pycodestyle==2.8.0
pytz==2022.1
requests==2.28.1
sqlparse==0.4.2
toml==0.10.2
urllib3==1.26.10
